Add fixed network mode for environments
Introduce a fixed network mode for environments. Specifically,
when this mode is activated, in the environment creation dialog
user is no longer prompted for a network and instead a network
previously assigned to the current project is used.

Network is assigned to project using 'murano_network' project metadata
key with network ID as the value.

This behavior is disabled by default and could be enabled by adding:

ENV_FIXED_NETWORK = yes

to the Horizon configuration.

Change-Id: Ic464e1560cc89fdc7615d51e67beb0f50e11f24c