Modify instance naming schema
When creating a new instance, generate its name based on
tenant name, application name and a region, therefore not using
a complicated composite structure used in Murano (murano+hashed),
and just UI generated name is displayed.

Also, the new schema takes into account names of the already
created instances to avoid duplication. For now, the 'hostname'
field for applications are still readonly, that would be changed
as validation is implemented.

Signed-off-by: Roman Bogorodskiy <rbogorodskiy@mirantis.com>
Change-Id: I7bd0758cf4b04a01531c788ebedc06a5c98fe523
Signed-off-by: Leonid Bogdanov <lbogdanov@mirantis.com>