Allow specification of only neutron net without subnet
in addition to existing approach where we are allowed to choose only
network with specific subnet.
It is useful when we have lots of small subnets and it is possible
to exceed amount of available IP addresses in it. In this case, we
will allow user to specify only network and delegate subnet choice
to server.

Change-Id: I23c21fe65552c0def527c98a5fa6a4eae56a6a7a